requirements.txt
Flask==3.0.3
ipython==8.12.3
protobuf==5.27.3
python-dotenv==1.0.1